fix(frontend): gate MagicBlock claim room as devnet demo
